Adress the cluade comments.
